requirements.txt
flask
gunicorn
requests==2.32.3
paho-mqtt==1.6.1
